Fix issues with the minify command
When running the concat task grunt would append the minified js to the
end of the minified file. Switching the position of the filters in the
uglify config fixed that issue.

Also added a js task that can be runt with `grunt js` to do the js tasks
but not css.