Make true the default state of showing tooltips and overviews
The default state of whether or not to show tooltips/overviews should
be to show them, so now we always save a meta value or either 0 or 1.

Cleaned up the user meta display and save functions a little as well.
Used the WP_User object that is passed to the action instead of
`get_user_meta()`, removed conditionals around the `checked()` call,
and switch save function to use `update_user_meta()` since
`update_usermeta` is deprecated.